quiz1/requirements.txt
flask
pandas
pymssql
sqlalchemy
azure.storage.blob
pillow